I am rm rfing everything I just did because that was a horrible idea
my original plan was for UICd and UIC clients to keep track of each
other's support for commands and parameters so that they wouldn't waste
their processing power and bandwidth to send things that can't be
handled. I decided that this "feature" has a larger imprint on memory
usage and code messiness than it is worth. commands and all parameters
will be sent to between all clients and servers, and anything that a
server or client does not know how to handle properly will be ignored
completely.